[Redesign] Separate fee computation
